+++ shared/src/androidMain/kotlin/org/hyperskill/app/step_quiz_hints/injection/PlatformStepQuizHintsComponentImpl.kt
@@ -2,9 +2,7 @@ package org.hyperskill.app.step_quiz_hints.injection
 
 import org.hyperskill.app.core.injection.AppGraph
 import org.hyperskill.app.core.injection.ReduxViewModelFactory
-import org.hyperskill.app.core.presentation.transformState
 import org.hyperskill.app.step.domain.model.Step
-import org.hyperskill.app.step_quiz_hints.mapper.StepQuizHintsViewStateMapper
 import org.hyperskill.app.step_quiz_hints.presentation.StepQuizHintsViewModel
 import ru.nobird.app.presentation.redux.container.wrapWithViewContainer
 
@@ -21,9 +19,7 @@ class PlatformStepQuizHintsComponentImpl(
                 StepQuizHintsViewModel::class.java to
                     {
                         StepQuizHintsViewModel(
-                            reduxViewContainer = stepQuizHintComponent.stepQuizHintsFeature
-                                .transformState(StepQuizHintsViewStateMapper::mapState)
-                                .wrapWithViewContainer(),
+                            reduxViewContainer = stepQuizHintComponent.stepQuizHintsFeature.wrapWithViewContainer(),
                             step = step
                         )
                     }
