+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz_hints/injection/StepQuizHintsComponent.kt
@@ -1,8 +1,9 @@
 package org.hyperskill.app.step_quiz_hints.injection
 
 import org.hyperskill.app.step_quiz_hints.presentation.StepQuizHintsFeature
+import org.hyperskill.app.step_quiz_hints.view.model.StepQuizHintsViewState
 import ru.nobird.app.presentation.redux.feature.Feature
 
 interface StepQuizHintsComponent {
-    val stepQuizHintsFeature: Feature<StepQuizHintsFeature.State, StepQuizHintsFeature.Message, StepQuizHintsFeature.Action>
+    val stepQuizHintsFeature: Feature<StepQuizHintsViewState, StepQuizHintsFeature.Message, StepQuizHintsFeature.Action>
 }
