+++ shared/src/commonTest/kotlin/org/hyperskill/StepQuizHintsViewStateMapperTest.kt
@@ -1,10 +1,10 @@
 package org.hyperskill
 
-import org.hyperskill.app.step_quiz_hints.mapper.StepQuizHintsViewStateMapper
-import org.hyperskill.app.step_quiz_hints.model.StepQuizHintsViewState
-import org.hyperskill.app.step_quiz_hints.presentation.StepQuizHintsFeature
-import org.junit.Test
+import kotlin.test.Test
 import kotlin.test.assertIs
+import org.hyperskill.app.step_quiz_hints.presentation.StepQuizHintsFeature
+import org.hyperskill.app.step_quiz_hints.view.mapper.StepQuizHintsViewStateMapper
+import org.hyperskill.app.step_quiz_hints.view.model.StepQuizHintsViewState
 
 class StepQuizHintsViewStateMapperTest {
     @Test
